requirements.txt
argcomplete==2.0.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
attrs==22.2.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
bandit==1.7.4 ; python_version >= "3.8" and python_full_version <= "3.10.0"
black==23.1a1 ; python_version >= "3.8" and python_full_version <= "3.10.0"
click==8.1.3 ; python_version >= "3.8" and python_full_version <= "3.10.0"
colorama==0.4.6 ; python_version >= "3.8" and python_full_version <= "3.10.0" and platform_system == "Windows" or python_version >= "3.8" and python_full_version <= "3.10.0" and sys_platform == "win32"
colorlog==6.7.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
dateutils==0.6.12 ; python_version >= "3.8" and python_full_version <= "3.10.0"
distlib==0.3.6 ; python_version >= "3.8" and python_full_version <= "3.10.0"
docutils==0.17.1 ; python_version >= "3.8" and python_full_version <= "3.10.0"
filelock==3.9.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
flake8-annotations==2.9.1 ; python_version >= "3.8" and python_full_version <= "3.10.0"
flake8-bandit==4.1.1 ; python_version >= "3.8" and python_full_version <= "3.10.0"
flake8-bugbear==22.12.6 ; python_version >= "3.8" and python_full_version <= "3.10.0"
flake8-docstrings==1.6.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
flake8-rst-docstrings==0.3.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
flake8==5.0.4 ; python_version >= "3.8" and python_full_version <= "3.10.0"
gitdb==4.0.10 ; python_version >= "3.8" and python_full_version <= "3.10.0"
gitpython==3.1.30 ; python_version >= "3.8" and python_full_version <= "3.10.0"
joblib==1.2.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
mccabe==0.7.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
mypy-extensions==0.4.3 ; python_version >= "3.8" and python_full_version <= "3.10.0"
nox==2022.11.21 ; python_version >= "3.8" and python_full_version <= "3.10.0"
numpy==1.24.1 ; python_version >= "3.8" and python_full_version <= "3.10.0"
packaging==21.3 ; python_version >= "3.8" and python_full_version <= "3.10.0"
pandas==1.5.2 ; python_version >= "3.8" and python_full_version <= "3.10.0"
pathspec==0.10.3 ; python_version >= "3.8" and python_full_version <= "3.10.0"
pbr==5.11.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
platformdirs==2.6.2 ; python_version >= "3.8" and python_full_version <= "3.10.0"
py4j==0.10.9.5 ; python_version >= "3.8" and python_full_version <= "3.10.0"
pycodestyle==2.9.1 ; python_version >= "3.8" and python_full_version <= "3.10.0"
pydocstyle==6.2.3 ; python_version >= "3.8" and python_full_version <= "3.10.0"
pyflakes==2.5.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
pygments==2.14.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
pyodbc==4.0.35 ; python_version >= "3.8" and python_full_version <= "3.10.0"
pyparsing==3.0.9 ; python_version >= "3.8" and python_full_version <= "3.10.0"
pyspark==3.3.1 ; python_version >= "3.8" and python_full_version <= "3.10.0"
python-dateutil==2.8.2 ; python_version >= "3.8" and python_full_version <= "3.10.0"
python-dotenv==0.19.2 ; python_version >= "3.8" and python_full_version <= "3.10.0"
pytz==2022.7 ; python_version >= "3.8" and python_full_version <= "3.10.0"
pyyaml==6.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
restructuredtext-lint==1.4.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
scikit-learn==1.2.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
scipy==1.10.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
six==1.16.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
smmap==5.0.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
snowballstemmer==2.2.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
stevedore==3.5.2 ; python_version >= "3.8" and python_full_version <= "3.10.0"
threadpoolctl==3.1.0 ; python_version >= "3.8" and python_full_version <= "3.10.0"
tomli==2.0.1 ; python_version >= "3.8" and python_full_version <= "3.10.0"
typing-extensions==4.4.0 ; python_version >= "3.8" and python_version < "3.10"
virtualenv==20.16.2 ; python_version >= "3.8" and python_full_version <= "3.10.0"
